Multiobjetivo (metas, pesos, etc.)
This linear optimization projects count with several goal functions, which means their goal is to optimize several characteristics of the result at a time such as the least distance, the equity, the safety, etc. in order to satisfy the demand of different cities (nodes) of a graph (connected cities of the country) once there has happened an unexpected phenomenon or catastrophe.